+++ common/src/main/java/com/www/common/pojo/dto/response/ResponseDTO.java
@@ -24,7 +24,7 @@ public class ResponseDTO<T> implements Serializable {
     /** 当前页数 **/
     private Integer pageNum;
     /** 页面条数 **/
-    private Integer pageSize;
+    private Long pageSize;
     /** 总数 **/
     private Long totalNum;
     /** 响应消息 **/
